created EntitiesField component
